rename VesselV2 types
