+++ libs/api-types/src/download.ts
@@ -1,0 +1,6 @@
+export type DownloadRateLimit = {
+  remaining: number
+  limit: number
+  reset?: string
+  retryAfter: number
+}
