+++ libs/api-types/src/index.ts
@@ -1,5 +1,6 @@
 export * from './datasets'
 export * from './dataviews'
+export * from './download'
 export * from './downloadActivity'
 export * from './events'
 export * from './geometries'
@@ -12,7 +13,6 @@ export * from './stats'
 export * from './user-applications'
 export * from './user'
 export * from './vessel'
-// TODO merge with vessel once new api is fully integrated
 export * from './identity-vessel'
 export * from './vesselGroups'
 export * from './workspaces'
